+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Updated
+- Insurance data for 2023 with new crop Wheat [FD-517](https://opensource.ncsa.illinois.edu/jira/browse/FD-517)]
+
 ## [1.6.1] - 2022-02-03
 
 ### Added